tcsetattr et tcgetattr savent manipuler le mode canonique et l'echo d'un tty. Mais rien de plus !

+++ libc/ioctl.c
@@ -1,9 +1,10 @@
+#include <types.h>
 #include <unistd.h>
 #include <sys/ioctl.h>
 #include <sys/syscall.h>
 
 int ioctl(int d, unsigned int request, void* data)
 {
-	syscall(SYS_IOCTL, d, request, data);
+	syscall(SYS_IOCTL, d, request, (uint32_t) data);
 	return 0;
 }
